CHOC-19: xrootd: Consolidate both xrootd ports under a common service.
Initially we were using the https port via the nginx ingress, and the
root port via a dedicated service. But this results in issues as the
hostname and certificate need to match for both. Therefore we remove the
ingress to run xrootd on a dedicated service with both https(443) and
root(1094) ports.