CHOC-146: set singleuser.cloudMetadata.blockWithIptables: false
The default is set to true, which results in jupyterhub_config.py
requesting NET_ADMIN capabilities for the container, which is blocked by
the cluster policies.
According to https://z2jh.jupyter.org/en/latest/administrator/security.html#block-cloud-metadata-api-with-a-networkpolicy-enforced-by-a-networkpolicy-controller
especially when using Cilium or Calico (we're using Cilium), we can just
rely on the NetworkPolicy controller of our Kubernetes cluster. The
recommendation is to set this to false.